HomeAutomationServer/requirements.txt
setuptools==40.8.0
pip==19.2.3
socketio
flask
paho-mqtt
jsonify
influxdb